+++ src/main/java/com/example/extra/domain/company/service/impl/CompanyServiceImpl.java
@@ -13,9 +13,9 @@ import com.example.extra.domain.company.service.CompanyService;
 import com.example.extra.domain.member.exception.MemberErrorCode;
 import com.example.extra.domain.member.exception.MemberException;
 import com.example.extra.domain.member.repository.MemberRepository;
+import com.example.extra.domain.refreshtoken.repository.RefreshTokenRepository;
+import com.example.extra.domain.refreshtoken.token.RefreshToken;
 import com.example.extra.global.security.JwtUtil;
-import com.example.extra.global.security.repository.RefreshTokenRepository;
-import com.example.extra.global.security.token.RefreshToken;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.security.crypto.password.PasswordEncoder;
