+++ src/main/java/com/example/extra/domain/member/service/impl/MemberServiceImpl.java
@@ -13,14 +13,14 @@ import com.example.extra.domain.member.exception.MemberException;
 import com.example.extra.domain.member.mapper.entity.MemberEntityMapper;
 import com.example.extra.domain.member.repository.MemberRepository;
 import com.example.extra.domain.member.service.MemberService;
+import com.example.extra.domain.refreshtoken.repository.RefreshTokenRepository;
+import com.example.extra.domain.refreshtoken.token.RefreshToken;
 import com.example.extra.domain.tattoo.dto.service.request.TattooCreateServiceRequestDto;
 import com.example.extra.domain.tattoo.entity.Tattoo;
 import com.example.extra.domain.tattoo.mapper.entity.TattooEntityMapper;
 import com.example.extra.domain.tattoo.repository.TattooRepository;
 import com.example.extra.global.security.JwtUtil;
 import com.example.extra.global.security.UserDetailsImpl;
-import com.example.extra.global.security.repository.RefreshTokenRepository;
-import com.example.extra.global.security.token.RefreshToken;
 import jakarta.servlet.ServletException;
 import jakarta.servlet.http.HttpServletRequest;
 import java.io.IOException;
