+++ src/main/java/com/example/extra/domain/refreshtoken/exception/TokenException.java
@@ -1,4 +1,4 @@
-package com.example.extra.global.security.exception;
+package com.example.extra.domain.refreshtoken.exception;
 
 import com.example.extra.global.exception.CustomException;
 import lombok.Getter;
