+++ src/main/java/com/example/extra/domain/refreshtoken/repository/RefreshTokenRepository.java
@@ -1,6 +1,6 @@
-package com.example.extra.global.security.repository;
+package com.example.extra.domain.refreshtoken.repository;
 
-import com.example.extra.global.security.token.RefreshToken;
+import com.example.extra.domain.refreshtoken.token.RefreshToken;
 import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Repository;
